Code Review：16-bit 寄存器地址支持 + 协议与代码规范修复
- bsp_i2c2: 寄存器地址 uint8_t → uint16_t，WritePhase 发送高低两字节
- app_motor: reg 参数同步改为 uint16_t
- app_sample: 去掉 (uint8_t) 强转，寄存器地址全程 uint16_t
- app_protocol: 0x20/0x21/0x22 数据段适配 16-bit 寄存器地址；
  BULK_* 宏移到文件顶部；删除未使用 Proto_CmdHandler_t；
  新增 _Static_assert 保护 dataLen；批量读上限改为 32130（255×126）
- protocol.MD: 同步更新 0x20/0x21/0x22 数据段格式及批量读上限
- CLAUDE.md: 路线图 4.1~4.4 标记完成，补充 16-bit 寄存器地址约束规则

+++ App/Inc/app_motor.h
@@ -15,7 +15,7 @@
 /*                               API                                        */
 /*--------------------------------------------------------------------------*/
 
-ErrorStatus App_Motor_ReadReg(uint8_t reg, uint16_t *pVal);
-ErrorStatus App_Motor_WriteReg(uint8_t reg, uint16_t val);
+ErrorStatus App_Motor_ReadReg(uint16_t reg, uint16_t *pVal);
+ErrorStatus App_Motor_WriteReg(uint16_t reg, uint16_t val);
 
 #endif /* __APP_MOTOR_H */
